List item render fix, group multiple click fix, list item styling

+++ js/index2.js
@@ -3,6 +3,7 @@
 const add1 = document.querySelector('#submit');
 const add2 = document.querySelector('#addTask');
 let foundObject = '';
+let groupStorageTasks = '';
 
 
 let groupStorage = JSON.parse(localStorage.getItem('Groups'));
@@ -40,10 +41,11 @@ function renderGroups() {
 }
 
 
-
 function renderItemList(groupIdNumber) {
   let groupHeader = document.querySelector('#groupHeader');
+  let itemList = document.querySelector('#itemList');
   let itemGroupHTML = '';
+
   // Grabbing header and creating empty string for header
 
   foundObject = groupStorage.tasks.find((object) => {
@@ -51,53 +53,50 @@ function renderItemList(groupIdNumber) {
   })
   // Returning the tasks with matching ID
 
-  groupHeader.innerHTML += `${foundObject.groupName}`;
+  groupHeader.innerHTML = `${foundObject.groupName}`;
   // Displaying group name
 
-  let itemList = document.querySelector('#itemList');
-  itemList.innerHTML = foundObject.tasks;
-  // Grabbing itemList from DOM, setting found
-
-  // if else to prevent {object} displaying
-  
-  // Setting the task list equal to array of tasks within group
-
-  addItemsToList();
+  for (let i=0; i < foundObject.tasks.length; i++) {
+    itemGroupHTML += `<br><li><input type='radio'>${foundObject.tasks[i].name}`
+  }
+  itemList.innerHTML = itemGroupHTML;
+  // Finding and inserting tasks from HTML
 }
 
 
 function addItemsToList() {
   let newTaskInput = document.querySelector('#newTask').value;
-  let storedTask = new Task(newTaskInput);
-  // Grabbing newTask input and using constructor
 
-  foundObject.tasks.push(storedTask);
-  // Pushing new task group into global foundObject variable L5
+  if (newTaskInput == '') {
+    window.alert('Please ender a valid task name');
+  } else {
 
-  groupStorage.tasks = groupStorage.tasks.filter((item) => foundObject.id != item.id);
-  groupStorage.tasks.push(foundObject);
-  // Replacing the groupStorage tasks with updated list
-  
-  localStorage.clear('Groups');
-  localStorage.setItem(groupStorage.groupName, JSON.stringify(groupStorage));
-  // Clearing local storage and replacing it with new Groups,groupstorage
-  
-  renderListItems(groupStorage.tasks)
+    let storedTask = new Task(newTaskInput);
+    // Grabbing newTask input and using constructor
+
+    foundObject.tasks.push(storedTask);
+    // Pushing new task group into global foundObject variable L5
+
+    groupStorage.tasks = groupStorage.tasks.filter((item) => foundObject.id != item.id);
+    groupStorage.tasks.push(foundObject);
+    // Replacing the groupStorage tasks with updated list
+    
+    localStorage.clear('Groups');
+    localStorage.setItem(groupStorage.groupName, JSON.stringify(groupStorage));
+    // Clearing local storage and replacing it with new Groups,groupstorage
+    
+    renderNewListItems(groupStorage.tasks)
+  }
 }
 
-function renderListItems(groupStorageTasks) {
+function renderNewListItems(groupStorageTasks) {
   let itemList = document.querySelector('#itemList');
   let itemListHTML = '';
   let groupHeader = document.querySelector('#groupHeader').innerHTML
-  
-  // console.log(groupStorageTasks)
-  // console.log(groupStorage);
-
 
   foundItem = groupStorageTasks.find((tasks) => {
     return tasks.groupName === groupHeader;
   })
-  
   
   for (let i=0; i < foundItem.tasks.length; i++) {
     itemListHTML += `<li><input type='radio'>${foundItem.tasks[i].name}`
@@ -109,35 +108,6 @@ function renderListItems(groupStorageTasks) {
 
 
 
-// function renderGroups() {
-//   let groupList = document.querySelector('#groupList');
-//   let groupStorageHTML = '';
-//   for (let i=0; i < groupStorage.tasks.length; i++) {
-//     groupStorageHTML += `<br><li onclick='renderItemList("${groupStorage.tasks[i].id}")' id="${groupStorage.tasks[i].id}">${groupStorage.tasks[i].groupName}</li>`;
-//   }
-//   groupList.innerHTML = groupStorageHTML;
-// }
-
-
-
 add1.addEventListener('click', addGroup);
 add2.addEventListener('click', addItemsToList);
 
-
-// REST
-// Use the routes and the http method to communicate with the server
-
-/*
-* GET /candy - All the candy
-* GET /candy/snickers - for that one candy
-* GET /candy/type/chocolate - for all chocolate candy
-*
-* POST /candy/cinnamon-bears - create new cinnamon bears candy
-* /customers
-*
-* GET /customers - all customers
-* GET /customers/2343352345 - that one customer
-* GET candystore.com/customers/pending-orders - 
-*
-*
-*/
